# cmd/api/server/openapi.yaml
openapi: 3.1.0
info:
  title: Code Search API
  description: |
    A self-hosted code search and replace platform with support for batch operations across repositories.

    ## Features
    - Fast indexed search across all repositories (powered by Zoekt)
    - CLI for bulk operations (search & replace with MR creation)
    - GitHub & GitLab integration (clone, sync, create PRs/MRs)

    ## Authentication
    API requests can be authenticated using a Bearer token in the Authorization header.
  version: 1.0.0
  contact:
    name: Code Search Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://codesearch.example.com
    description: Production server
...
paths:
...
  /api/v1/repos/by-id/{id}/tree:
    get:
      tags:
        - File Browsing
      summary: List directory contents
      description: |
        Returns the contents of a directory in a repository.
        Can be disabled via `CS_UI_DISABLE_BROWSE_API`.
      operationId: listTree
      parameters:
        - $ref: "#/components/parameters/RepoId"
        - name: path
          in: query
          description: Path within the repository (empty for root)
          schema:
            type: string
            default: ""
        - name: ref
          in: query
          description: Git ref (branch, tag, or commit). Defaults to repository's default branch.
          schema:
            type: string
      responses:
        "200":
          description: Directory listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreeResponse"
        "404":
          description: Repository or path not found
        "503":
          description: Repository not cloned locally (indexer unavailable in sharded mode)
  /api/v1/repos/by-id/{id}/blob:
    get:
      tags:
        - File Browsing
      summary: Get file contents
      description: |
        Returns the contents of a file in a repository.
        Can be disabled via `CS_UI_DISABLE_BROWSE_API`.
      operationId: getBlob
      parameters:
        - $ref: "#/components/parameters/RepoId"
        - name: path
          in: query
          required: true
          description: Path to the file within the repository
          schema:
            type: string
        - name: ref
          in: query
          description: Git ref (branch, tag, or commit). Defaults to repository's default branch.
          schema:
            type: string
      responses:
        "200":
          description: File contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlobResponse"
        "404":
          description: Repository or file not found
        "503":
          description: Repository not cloned locally
  /api/v1/repos/by-id/{id}/raw:
    get:
      tags:
        - File Browsing
      summary: Get raw file contents
      description: |
        Returns the raw contents of a file without JSON wrapping.
        Useful for downloading files directly.
        Can be disabled via `CS_UI_DISABLE_BROWSE_API`.
      operationId: getRawBlob
      parameters:
        - $ref: "#/components/parameters/RepoId"
        - name: path
          in: query
          required: true
          description: Path to the file within the repository
          schema:
            type: string
        - name: ref
          in: query
          description: Git ref (branch, tag, or commit). Defaults to repository's default branch.
          schema:
            type: string
      responses:
        "200":
          description: Raw file contents
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
            text/plain:
              schema:
                type: string
        "404":
          description: Repository or file not found
        "503":
          description: Repository not cloned locally
...
  /api/v1/replace/execute:
    post:
      tags:
        - Replace
      summary: Execute replacements
      description: |
        Execute search and replace across repositories.
        Creates merge/pull requests with the changes.

        **Read-only mode:** When repositories are configured in read-only mode
        (check `/api/v1/repos/status`), users must provide their own personal
        access token with write permissions to create merge requests.
      operationId: replaceExecute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaceExecuteRequest"
      responses:
        "202":
          description: Replace job queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplaceExecuteResponse"
        "400":
          description: Invalid request
...
  /api/v1/scip/repos/{id}/status:
    get:
      tags:
        - SCIP
      summary: Get SCIP index status
      description: Returns whether a repository has a SCIP index and statistics about it.
      operationId: getSCIPStatus
      parameters:
        - $ref: "#/components/parameters/RepoId"
      responses:
        "200":
          description: SCIP index status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SCIPStatusResponse"
        "404":
          description: Repository not found
  /api/v1/scip/repos/{id}/definition:
    post:
      tags:
        - SCIP
      summary: Go to definition
      description: |
        Find the definition of a symbol at a given position in a file.
        Returns the location where the symbol is defined, along with documentation if available.
        Requires a SCIP index for the repository.
      operationId: goToDefinition
      parameters:
        - $ref: "#/components/parameters/RepoId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoToDefinitionRequest"
      responses:
        "200":
          description: Definition location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoToDefinitionResponse"
        "400":
          description: Invalid request
        "404":
          description: Repository not found
  /api/v1/scip/repos/{id}/references:
    post:
      tags:
        - SCIP
      summary: Find references
      description: |
        Find all references to a symbol at a given position in a file.
        Returns all locations where the symbol is used, including the definition.
        Requires a SCIP index for the repository.
      operationId: findSCIPReferences
      parameters:
        - $ref: "#/components/parameters/RepoId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FindReferencesRequest"
      responses:
        "200":
          description: References found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindReferencesResponse"
        "400":
          description: Invalid request
        "404":
          description: Repository not found
  /api/v1/scip/repos/{id}/index:
    post:
      tags:
        - SCIP
      summary: Index repository with SCIP
      description: |
        Run SCIP indexing on a repository. This creates a precise code intelligence index
        that enables go-to-definition and find-references.

        **Prerequisites:**
        - The appropriate SCIP indexer must be installed (e.g., `scip-go` for Go)
        - The repository must be cloned locally

        **Supported languages:**
        - Go (`scip-go`)
        - TypeScript/JavaScript (`@sourcegraph/scip-typescript`)
        - Java (`scip-java`)
        - Rust (`rust-analyzer`)
        - Python (`scip-python`)
        - PHP (`scip-php` via Composer)
      operationId: indexSCIP
      parameters:
        - $ref: "#/components/parameters/RepoId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IndexSCIPRequest"
      responses:
        "200":
          description: Indexing completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndexSCIPResponse"
        "400":
          description: Invalid request or indexer not available
        "404":
          description: Repository not found
        "500":
          description: Indexing failed

    delete:
      tags:
        - SCIP
      summary: Clear SCIP index
      description: Remove the SCIP index for a repository.
      operationId: clearSCIPIndex
      parameters:
        - $ref: "#/components/parameters/RepoId"
      responses:
        "200":
          description: Index cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: SCIP index cleared
        "404":
          description: Repository not found
...
  /api/v1/scip/repos/{id}/symbols/search:
    post:
      tags:
        - SCIP
      summary: Search symbols in SCIP index
      description: Search for symbols by name in a repository's SCIP index.
      operationId: searchSCIPSymbols
      parameters:
        - $ref: "#/components/parameters/RepoId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchSymbolsRequest"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchSymbolsResponse"
        "400":
          description: Invalid request
        "404":
          description: Repository not found
  /api/v1/scip/repos/{id}/files:
    get:
      tags:
        - SCIP
      summary: List indexed files
      description: List all files that are included in the SCIP index for a repository.
      operationId: listSCIPFiles
      parameters:
        - $ref: "#/components/parameters/RepoId"
        - name: limit
          in: query
          description: Maximum number of files to return
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
      responses:
        "200":
          description: List of indexed files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListSCIPFilesResponse"
        "404":
          description: Repository not found

  # Scheduler Endpoints
...
components:
...
  parameters:
    RepoId:
      name: id
      in: path
      required: true
      description: Repository ID
      schema:
        type: integer
        format: int64
...
  schemas:
...
    TreeEntry:
      type: object
      properties:
        name:
          type: string
          description: File or directory name
          example: src
        path:
          type: string
          description: Full path within the repository
          example: src
        type:
          type: string
          enum: [file, dir]
          description: Entry type
          example: dir
        size:
          type: integer
          format: int64
          description: File size in bytes (only for files)
          example: 1234
    TreeResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/TreeEntry"
        path:
          type: string
          description: Current directory path
          example: src
        ref:
          type: string
          description: Git ref used
          example: main
    BlobResponse:
      type: object
      properties:
        content:
          type: string
          description: File content (empty for binary files)
        path:
          type: string
          description: File path
          example: src/main.go
        size:
          type: integer
          format: int64
          description: File size in bytes
          example: 1234
        language:
          type: string
          description: Detected programming language
          example: Go
        language_mode:
          type: string
          description: CodeMirror syntax highlighting mode
          example: go
        binary:
          type: boolean
          description: Whether the file is binary
          example: false
        ref:
          type: string
          description: Git ref used
          example: main
...
    ReplaceExecuteRequest:
      type: object
      required:
        - search_pattern
        - replace_with
        - matches
      properties:
        search_pattern:
          type: string
          description: Pattern to search for
          example: oldFunction
        replace_with:
          type: string
          description: Replacement text
          example: newFunction
        is_regex:
          type: boolean
          default: false
        case_sensitive:
          type: boolean
          default: false
        file_patterns:
          type: array
          items:
            type: string
          example: ["*.go", "*.ts"]
        branch_name:
          type: string
          description: Branch name for changes (auto-generated if not specified)
        mr_title:
          type: string
          description: Title for the merge/pull request
        mr_description:
          type: string
          description: Description for the merge/pull request
        user_tokens:
          type: object
          additionalProperties:
            type: string
          description: |
            Map of connection_name to personal access token.
            Required when repositories are in read-only mode.
        matches:
          type: array
          description: |
            Matches from the preview endpoint. Required - you must call the preview
            endpoint first to get matches, then pass them here. The replace job uses
            these directly without re-searching. A merge request is always created -
            changes are never committed directly to the main branch.
          items:
            $ref: "#/components/schemas/ReplaceMatch"
...
    ReplaceExecuteResponse:
      type: object
      properties:
        job_id:
          type: string
        status:
          type: string
          example: pending
        message:
          type: string
          example: Replace job queued

    # Symbol Schemas
...
    SCIPStatusResponse:
      type: object
      properties:
        has_index:
          type: boolean
          description: Whether the repository has a SCIP index
        available_indexers:
          type: object
          description: Map of language to availability status
          additionalProperties:
            type: boolean
        stats:
          type: object
          description: Statistics about the index (only present if has_index is true)
          properties:
            total_files:
              type: integer
            total_symbols:
              type: integer
            total_occurrences:
              type: integer
    GoToDefinitionRequest:
      type: object
      required:
        - filePath
        - line
        - column
      properties:
        filePath:
          type: string
          description: Path to the file within the repository
          example: internal/search/search.go
        line:
          type: integer
          description: Line number (1-indexed, matching editor display)
          example: 42
        column:
          type: integer
          description: Column number (0-indexed)
          example: 10
    GoToDefinitionResponse:
      type: object
      properties:
        found:
          type: boolean
          description: Whether a definition was found
        symbol:
          type: string
          description: The SCIP symbol identifier
        definition:
          $ref: "#/components/schemas/SCIPOccurrence"
        info:
          $ref: "#/components/schemas/SCIPSymbolInfo"
        external:
          type: boolean
          description: Whether the definition is in an external package
    FindReferencesRequest:
      type: object
      required:
        - filePath
        - line
        - column
      properties:
        filePath:
          type: string
          description: Path to the file within the repository
          example: internal/search/search.go
        line:
          type: integer
          description: Line number (1-indexed, matching editor display)
          example: 42
        column:
          type: integer
          description: Column number (0-indexed)
          example: 10
        limit:
          type: integer
          description: Maximum number of references to return
          default: 100
    FindReferencesResponse:
      type: object
      properties:
        found:
          type: boolean
          description: Whether any references were found
        symbol:
          type: string
          description: The SCIP symbol identifier
        definition:
          $ref: "#/components/schemas/SCIPOccurrence"
        references:
          type: array
          items:
            $ref: "#/components/schemas/SCIPOccurrence"
        totalCount:
          type: integer
          description: Total number of references found
...
    IndexSCIPRequest:
      type: object
      properties:
        language:
          type: string
          description: Programming language (auto-detected if not specified)
          enum: [go, typescript, javascript, java, rust, python, php]
    IndexSCIPResponse:
      type: object
      properties:
        success:
          type: boolean
        language:
          type: string
          description: The language that was indexed
        duration:
          type: string
          description: Time taken to index
          example: "2.5s"
        files:
          type: integer
          description: Number of files indexed
        symbols:
          type: integer
          description: Number of symbols found
        occurrences:
          type: integer
          description: Number of symbol occurrences
        error:
          type: string
          description: Error message if indexing failed
        output:
          type: string
          description: Raw output from the indexer
    SearchSymbolsRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: Symbol name pattern to search for
          example: HandleRequest
        limit:
          type: integer
          description: Maximum number of results
          default: 100
    SearchSymbolsResponse:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              symbol:
                type: string
              name:
                type: string
              kind:
                type: string
              filePath:
                type: string
              line:
                type: integer
        count:
          type: integer
    ListSCIPFilesResponse:
      type: object
      properties:
        files:
          type: array
          items:
            type: string
          description: List of file paths in the SCIP index
        count:
          type: integer
          description: Number of files returned

    # UI Settings Schema
...